+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Version 3.0.0 - 2022-06-14
+### Changed
+- Refactored scripts
+  - Renamed a lot of variables ([PLEASE see docs](README.md#environment-variables))
+  - Updated logic used to select the OpenVPN configuration file
+  - Switched to `nftables`
+- Updated to Alpine 3.16
+- Fixed outdated proxy configuration files
+
 ## Version 2.1.0 - 2022-03-06
 ### Added
 - `VPN_CONFIG_PATTERN` environment variable.